onupdate showonlist column added in fields table

+++ administrator/sql/install.mysql.utf8.sql
@@ -15,7 +15,7 @@ CREATE TABLE IF NOT EXISTS `#__tjfields_fields` (
 	`filterable` tinyint(1) NOT NULL DEFAULT '0' COMMENT '0 - For not filterable field. 1 for filterable field',
 	`client` varchar(255) NOT NULL,
 	`group_id` int(11) NOT NULL,
-	`showonlist` int(11) NOT NULL,
+	`showonlist` tinyint(1) NOT NULL,
 	`params` varchar(500),
 	PRIMARY KEY (`id`)
 )   DEFAULT CHARSET=utf8 AUTO_INCREMENT=1 ;
